feat[checkDraw method]
a method to check if all spaces in a tic-tac-toe matrix
have been taken by binary value of X or O.if no
spaces availbe (no null values in the board array)
checkDraw returns true and no further moves can be
done, game is over.
It is not necessary to check if there is winning
position in the game for either player from witin this method
since checkWinningPositions will determine that and
will stop game play at that moment.